refactor(extensions): move Bundler + ModuleLexer + helper to src/extensions/bundler/, rename ext-esbuild → ext-bundler-esbuild
- src/extensions/bundler/{bundler,module-lexer,module-lexer.test,helper,index}.ts (new layout)
- src/extensions/bundler.ts (old top-level helper) moved to src/extensions/bundler/helper.ts
- Barrel splits the 22 interfaces (export type) from the Loader union (also export type)
  and the helper values build/getBundler (export). Extended beyond the plan template
  to also export context/stop/transform values and the 6 esbuild-flavoured type
  aliases (Plugin/PluginBuild/Message/ResolveResult/BuildOptions/BuildResult) that
  the old top-level bundler.ts forwarded, so existing consumers keep compiling.
- src/extensions/interfaces/index.ts shim re-exports from ../bundler/
- extensions/ext-bundler-esbuild (@veryfront/ext-bundler-esbuild)
- Root deno.json workspace + ./extensions/bundler export
- Consumer stragglers (builtin-extensions.ts, runtime name, init scaffold deps, tests) updated
