+++ deno.json
@@ -9,8 +9,8 @@
     "./extensions/ext-ai-openai",
     "./extensions/ext-auth-jwt",
     "./extensions/ext-babel",
+    "./extensions/ext-bundler-esbuild",
     "./extensions/ext-cache-redis",
-    "./extensions/ext-esbuild",
     "./extensions/ext-mdx",
     "./extensions/ext-node-compat",
     "./extensions/ext-opentelemetry",
@@ -71,6 +71,7 @@
     "./extensions/interfaces": "./src/extensions/interfaces/index.ts",
     "./extensions/ai": "./src/extensions/ai/index.ts",
     "./extensions/auth": "./src/extensions/auth/index.ts",
+    "./extensions/bundler": "./src/extensions/bundler/index.ts",
     "./extensions/cache": "./src/extensions/cache/index.ts",
     "./testing": "./src/testing/index.ts",
     "./testing/assert": "./src/testing/assert.ts",
@@ -154,7 +155,7 @@
     "veryfront/testing": "./src/testing/index.ts",
     "veryfront/testing/assert": "./src/testing/assert.ts",
     "veryfront/testing/bdd": "./src/testing/bdd.ts",
-    "veryfront/extensions/bundler": "./src/extensions/bundler.ts",
+    "veryfront/extensions/bundler": "./src/extensions/bundler/index.ts",
     "veryfront/extensions/contracts": "./src/extensions/contracts.ts",
     "veryfront/extensions/interfaces": "./src/extensions/interfaces/index.ts",
     "jose": "npm:jose@5.9.6",
