+++ src/extensions/recommendations.ts
@@ -5,10 +5,10 @@
  */
 
 const recommendations = new Map<string, string>([
-  ["Bundler", "@veryfront/ext-esbuild"],
+  ["Bundler", "@veryfront/ext-bundler-esbuild"],
   // ModuleLexer ships in the same package as Bundler (es-module-lexer +
-  // esbuild are co-bundled into ext-esbuild).
-  ["ModuleLexer", "@veryfront/ext-esbuild"],
+  // esbuild are co-bundled into ext-bundler-esbuild).
+  ["ModuleLexer", "@veryfront/ext-bundler-esbuild"],
   ["CacheStore", "@veryfront/ext-cache-redis"],
   ["TokenCacheStore", "@veryfront/ext-cache-redis"],
   ["CSSProcessor", "@veryfront/ext-tailwind"],
