+++ tests/_helpers/contract-init.ts
@@ -2,7 +2,7 @@
  * Test-environment contract registration.
  *
  * Side-effect import: registers the default Bundler and ModuleLexer
- * implementations from `@veryfront/ext-esbuild` so integration tests that
+ * implementations from `@veryfront/ext-bundler-esbuild` so integration tests that
  * exercise bootstrap/build paths can resolve those contracts without each
  * test having to wire it up.
  *
@@ -11,7 +11,7 @@
  * @module
  */
 
-import { EsbuildBundler, EsModuleLexer } from "@veryfront/ext-esbuild";
+import { EsbuildBundler, EsModuleLexer } from "@veryfront/ext-bundler-esbuild";
 import { register, tryResolve } from "../../src/extensions/contracts.ts";
 
 if (!tryResolve("Bundler")) register("Bundler", new EsbuildBundler());
